move cask/cmd/list to cask/list
