+++ Library/Homebrew/cask/cmd.rb
@@ -13,7 +13,6 @@ require "cask/cmd/abstract_command"
 require "cask/cmd/audit"
 require "cask/cmd/fetch"
 require "cask/cmd/install"
-require "cask/cmd/list"
 require "cask/cmd/reinstall"
 require "cask/cmd/uninstall"
 require "cask/cmd/upgrade"
